Tidying
1) Functions which evaluate the derivatives of H have been placed in a different file. This is to enable easy switching between different numerical schemes.

2) Added a function find_widths_and_curvatures to find the widths and curvatures.
